Move StatsAggregator to its own file
Also, make the code more pep8 compliant and refactor it a bit to remove
some duplication.
