+++ rdbtools/__init__.py
@@ -1,6 +1,7 @@
 from rdbtools.parser import RdbCallback, RdbParser, DebugCallback
 from rdbtools.callbacks import JSONCallback, DiffCallback, ProtocolCallback
-from rdbtools.memprofiler import MemoryCallback, PrintAllKeys, StatsAggregator
+from rdbtools.memprofiler import MemoryCallback, PrintAllKeys
+from rdbtools.stats_aggregator import StatsAggregator
 
 __version__ = '0.1.6'
 VERSION = tuple(map(int, __version__.split('.')))
