+++ rdbtools/memprofiler.py
@@ -1,6 +1,5 @@
-from collections import defaultdict, namedtuple
+from collections import namedtuple
 import random
-import json
 
 from rdbtools.parser import RdbCallback
 from rdbtools.callbacks import encode_key
@@ -11,48 +10,7 @@ REDIS_SHARED_INTEGERS = 10000
 
 MemoryRecord = namedtuple('MemoryRecord', ['database', 'type', 'key', 'bytes', 'encoding','size', 'len_largest_element'])
 
-class StatsAggregator(object):
-    def __init__(self):
-        self.aggregates = defaultdict(lambda: defaultdict(int))
-        self.histograms = defaultdict(lambda: defaultdict(int))
-        self.scatters = defaultdict(list)
 
-    def next_record(self, record):
-        self.add_aggregate('database_memory', record.database, record.bytes)
-        self.add_aggregate('type_memory', record.type, record.bytes)
-        self.add_aggregate('encoding_memory', record.encoding, record.bytes)
-        
-        self.add_aggregate('type_count', record.type, 1)
-        self.add_aggregate('encoding_count', record.encoding, 1)
-    
-        self.add_histogram(record.type + "_length", record.size)
-        self.add_histogram(record.type + "_memory", (record.bytes/10) * 10)
-        
-        if record.type == 'list':
-            self.add_scatter('list_memory_by_length', record.bytes, record.size)
-        elif record.type == 'hash':
-            self.add_scatter('hash_memory_by_length', record.bytes, record.size)
-        elif record.type == 'set':
-            self.add_scatter('set_memory_by_length', record.bytes, record.size)
-        elif record.type == 'sortedset':
-            self.add_scatter('sortedset_memory_by_length', record.bytes, record.size)
-        elif record.type == 'string':
-            self.add_scatter('string_memory_by_length', record.bytes, record.size)
-        else:
-            raise Exception('Invalid data type %s' % record.type)
-
-    def add_aggregate(self, heading, subheading, metric):
-        self.aggregates[heading][subheading] += metric
-
-    def add_histogram(self, heading, metric):
-        self.histograms[heading][metric] += 1
-
-    def add_scatter(self, heading, x, y):
-        self.scatters[heading].append([x, y])
-
-    def get_json(self):
-        return json.dumps({"aggregates":self.aggregates, "scatters":self.scatters, "histograms":self.histograms})
-        
 class PrintAllKeys():
     def __init__(self, out):
         self._out = out
